FIX pull cache does not work with BSD date
docker already manages a cach of images so that pulling an image
we already have locally is fast. Hence, I do not see the point to
keep that code wich seems to prevent pulling an image more that once
a week.

Moreover, this code won't work on os x as it relies on BSD date and
GNU and BSD date are not so much friends :(

I propose we just get rid of this block of code.